common selfTop: take 'ind' as a 'size_t'
Pays down some cast noise from prior commits.

+++ lib/common/splines.c
@@ -886,7 +886,7 @@ static void selfBottom(edge_t *edges[], int ind, size_t cnt, double sizex,
     }
 }
 
-static void selfTop(edge_t *edges[], int ind, size_t cnt, double sizex,
+static void selfTop(edge_t *edges[], size_t ind, size_t cnt, double sizex,
                     double stepy, splineInfo *sinfo) {
     int sgn, point_pair;
     double hy, ty,  stepx, dx, dy, height;
@@ -1193,7 +1193,7 @@ void makeSelfEdge(edge_t *edges[], size_t ind, size_t cnt, double sizex,
 
 	/* handle L-R specially */
 	if ((ED_tail_port(e).side & RIGHT) || (ED_head_port(e).side & RIGHT)) {
-	    selfTop(edges, (int)ind, cnt, sizex, sizey, sinfo);
+	    selfTop(edges, ind, cnt, sizex, sizey, sinfo);
 	}
 	else {
 	    selfLeft(edges, ind, cnt, sizex, sizey, sinfo);
@@ -1202,7 +1202,7 @@ void makeSelfEdge(edge_t *edges[], size_t ind, size_t cnt, double sizex,
 
     /* self edge with both ports on top side */
     else if (ED_tail_port(e).side & TOP) {
-	selfTop(edges, (int)ind, cnt, sizex, sizey, sinfo);
+	selfTop(edges, ind, cnt, sizex, sizey, sinfo);
     }
     else if (ED_tail_port(e).side & BOTTOM) {
 	selfBottom(edges, (int)ind, cnt, sizex, sizey, sinfo);
